media-feed-service/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Flask-Migrate==3.1.0
python-jose==3.3.0
requests==2.26.0
Pillow==9.0.0
PyJWT==2.1.0
werkzeug==2.0.2
SQLAlchemy==1.4.36